requirements.txt
numpy
matplotlib
torch==1.11.0
torchvision==0.12.0
pillow
opencv-contrib-python